+++ DobieStation/application/application.pro
@@ -62,6 +62,7 @@ SOURCES += ../../src/qt/main.cpp \
     ../../src/core/iop/spu/spu_adpcm.cpp \
     ../../src/core/iop/spu/spu_envelope.cpp \
     ../../src/core/iop/spu/spu_interpolate.cpp \
+    ../../src/core/iop/spu/spu_reverb.cpp \
     ../../src/qt/emuthread.cpp \
     ../../src/core/tests/iop/alu.cpp \
     ../../src/core/ee/vif.cpp \
