+++ src/core/CMakeLists.txt
@@ -87,6 +87,7 @@ set(SOURCES
     iop/spu/spu_adpcm.cpp
     iop/spu/spu_envelope.cpp
     iop/spu/spu_interpolate.cpp
+    iop/spu/spu_reverb.cpp
     jitcommon/emitter64.cpp
     jitcommon/ir_block.cpp
     jitcommon/ir_instr.cpp
